+++ CHANGES.md
@@ -24,6 +24,8 @@
 * Remove `Linear::DSL.insert_task`. The canonical way to add steps is using the ADDS interface going through a normalizer.
   This is why there's a normalizer for `end` (or "terminus") now for consistency.
 * In Normalizer, `path.wirings` is now `activity.wirings`.
+* Remove `Helper::ClassMethods`, `Helper` is now the namespace to mix in your own macros (and ours, like `Output()`).
+* Move DSL structures like `OutputSemantic` to `Linear` namespace.
 
 # 0.5.0